Split logic into messenger, auth, and requester
